Adding Superman shows to MainPage
